python-barcode-scanner/requirements.txt
opencv-python==4.8.1.78
pyzbar==0.1.9
Pillow==10.0.1
numpy==1.24.3
python-barcode==0.15.1
qrcode[pil]==7.4.2
cryptography==41.0.7
requests==2.31.0
